[Fix] #51 - 뷰모델 의존성 주입 처리
